requirements.txt
# Meridian Bot Dependencies

# Telegram Bot
python-telegram-bot==20.7

# HTTP Client
aiohttp==3.9.1

# WebSocket
websockets==12.0

# Environment Variables
python-dotenv==1.0.0

# Database
# SQLite is built into Python

# Logging and Utilities
# Standard library components

# Optional: For enhanced features
# APScheduler==3.10.4  # For scheduled tasks
# pandas==2.1.4  # For data analysis
# matplotlib==3.8.2  # For charts (if adding visualization)
